# openapi.yaml
openapi: 3.0.3
info:
  title: Rabit HR Platform API
  description: |
    Complete HR Management Platform API with comprehensive features for:
    - Employee Management
    - Attendance & Leave Management
    - Payroll & Benefits
    - Performance Reviews
    - Training & Development
    - Recruitment
    - Document Management
    - Consultation Services
  version: 1.0.0
  contact:
    name: Rabit HR Support
    email: support@rabithr.com
    url: https://rabithr.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.rabithr.com
    description: Production server
...
paths:
...
  /api/auth/login:
    post:
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin@rabithr.com
                password:
                  type: string
                  format: password
                  example: Password123!
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, hr, manager, employee, consultant]
        fullName:
          type: string
        avatar:
          type: string
          format: uri
...
security:
  - bearerAuth: []